requirement.txt
flask
dotenv
mysql-connector-python
bcrypt
eventlet
gevent